--- js/random-function.js
@@ -1,9 +1,0 @@
-//генерация рамдомного ID элемента
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-export { getRandomInteger };
